refactor(cli): extract shared mesh-connect helper; migrate transient commands
New lib/connect.ts holds resolveTargetOrExit + preflightOrExit, extracted from spawn.ts
so every mesh-touching command resolves WHICH mesh + creds the same way (resolveMeshTarget)
and gets the same one-sentence preflight instead of a raw NATS Authorization Violation from
outside a project. preflightOrExit now takes optional probe creds (so the --creds escape
hatch probes with the creds it'll actually use). spawn.ts imports the helper; lib/transient.ts
(resolveConnect/openTransient) drops the cotalRoot()/loadSpaceAuth/isReachable cwd-walk for
the shared resolver — which migrates send (dm/msg/ask), history, channels, and personas
--running in one move, so they work from any directory.
