+++ implementations/cli/src/commands/spawn.ts
@@ -14,20 +14,17 @@ import {
   mintCreds,
   newIdentity,
   parseShareSelection,
-  probeConnect,
   provisionAgent,
   registry,
-  removeMesh,
   resolveMeshTarget,
   CotalEndpoint,
   type AgentDef,
   type CompletionResult,
   type Connector,
-  type MeshTarget,
   type SpaceAuth,
 } from "@cotal-ai/core";
 import { c } from "../ui.js";
-import { pruneStaleMeshes } from "../lib/meshes.js";
+import { preflightOrExit, resolveTargetOrExit } from "../lib/connect.js";
 import { listPersonas } from "../lib/personas.js";
 
 /** Completion for `cotal spawn` — `--space <TAB>` lists the running meshes, and the first positional
@@ -115,78 +112,6 @@ async function uniqueMeshName(
   }
 }
 
-/** Resolve the mesh this spawn targets, exiting with one human sentence on an unresolved/ambiguous
- *  registry rather than a stack trace. Prunes dead registry entries first so a crashed mesh doesn't
- *  block a bare spawn or get offered by `--space`. */
-async function resolveTargetOrExit(flags: { server?: string; space?: string }): Promise<MeshTarget> {
-  await pruneStaleMeshes();
-  try {
-    return resolveMeshTarget(process.cwd(), flags);
-  } catch (e) {
-    console.error(c.red(`✗ ${(e as Error).message}`));
-    process.exit(1);
-  }
-}
-
-/** Confirm the resolved mesh is actually up and accepts our creds — replaces the raw NATS
- *  "Authorization Violation" trace with one sentence, and prunes the entry if the broker is gone. */
-async function preflightOrExit(target: MeshTarget): Promise<void> {
-  const creds = target.auth ? await mintCreds(target.auth, newIdentity(), "manager") : undefined;
-  const probe = await probeConnect(target.server, creds ? { creds } : {});
-  if (probe.ok) return;
-  // A target whose server + mode came from a registry record owns that record, so a definitive
-  // failure prunes the stale entry. `local-recorded` (a local project matched to an entry by root)
-  // is registry-owned for pruning even though the success line treats it as a quiet local target.
-  const fromRegistry =
-    target.source === "registry" ||
-    target.source === "current" ||
-    target.source === "flag-space" ||
-    target.source === "local-recorded";
-  if (probe.reason === "unreachable") {
-    if (fromRegistry) removeMesh(target.space); // the broker is gone — drop the stale entry
-    console.error(
-      c.red(
-        `✗ no mesh running at ${target.server}${fromRegistry ? " (stale registry entry — removed)" : ""} — run \`cotal up\``,
-      ),
-    );
-  } else if (fromRegistry && target.auth) {
-    // We DID present creds minted from the registry's own trust material and the broker rejected
-    // them — so the entry now points at a DIFFERENT broker on that port. The fix isn't "spawn from
-    // the root" (we just used it); it's that the registry entry is stale. Drop it and say so.
-    removeMesh(target.space);
-    console.error(
-      c.red(
-        `✗ mesh "${target.space}" at ${target.server} no longer matches its registry entry (credentials rejected — port reused?) — re-run \`cotal up\` from ${target.root}, or \`cotal meshes\` to see what's live`,
-      ),
-    );
-  } else if (fromRegistry) {
-    // An OPEN-mode registry target probed credlessly, but the broker now wants auth — the recorded
-    // open mesh is gone (its port was reused by an auth broker). Same stale class; drop the entry.
-    removeMesh(target.space);
-    console.error(
-      c.red(
-        `✗ open mesh "${target.space}" at ${target.server} no longer matches its registry entry (broker now requires auth — port reused?) — re-run \`cotal up\` from ${target.root}, or \`cotal meshes\` to see what's live`,
-      ),
-    );
-  } else if (target.auth) {
-    // Creds were presented and rejected, but the target is a local project / --server we don't own
-    // in the registry. A different mesh is likely on that port — the user owns the diagnosis, so we
-    // don't prune; we point at the likely cause. "Can't provide" would be wrong: the folder DID.
-    console.error(
-      c.red(
-        `✗ credentials for "${target.space}" were rejected at ${target.server} — a different mesh may be running there. Run \`cotal meshes\` to check, or \`cotal up\` here to start yours`,
-      ),
-    );
-  } else {
-    // No creds to present (open, non-registry) and the broker wants auth.
-    console.error(
-      c.red(
-        `✗ broker at ${target.server} requires auth, but this mesh is open (no trust material) — use \`--space <name>\` for an auth mesh, or run \`cotal up\` here without \`--open\``,
-      ),
-    );
-  }
-  process.exit(1);
-}
 
 export async function spawn(argv: string[]): Promise<void> {
   const { values, positionals } = parseArgs({
